backend/pyproject.toml
[project]
name = "resumeai-api"
version = "0.1.0"
description = "Resume AI Analyzer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt[crypto]>=5.5.0",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "gunicorn>=21.2.0",
    "mysqlclient>=2.2.3",
    "django-redis>=5.4.0",
    "django-cors-headers>=4.3.1",
    "pillow>=11.2.1",
    # Resume parsing dependencies
    "pypdf2>=3.0.0",
    "docx2txt>=0.8",
    "nltk>=3.8.1",
    "spacy>=3.7.0",
    "celery>=5.5.1",
    "django-celery-results>=2.6.0",
    "pydantic[email]>=2.11.3",
    "pymongo>=4.12.0",
]
